* Fix a couple of minor compiler nits EPIC4-1.2.1!

+++ source/window.c
@@ -1,4 +1,4 @@
-/* $EPIC: window.c,v 1.83 2003/12/02 07:24:55 jnelson Exp $ */
+/* $EPIC: window.c,v 1.84 2003/12/02 21:02:11 jnelson Exp $ */
 /*
  * window.c: Handles the organzation of the logical viewports (``windows'')
  * for irc.  This includes keeping track of what windows are open, where they
@@ -4699,11 +4699,11 @@ BUILT_IN_COMMAND(windowcmd)
 		{
 			if (!my_strnicmp(arg, options[i].command, len))
 			{
-				winref = window ? window->refnum : -1;
+				winref = window ? (int)window->refnum : -1;
 				window = options[i].func(window, &args); 
 				nargs++;
 				do_hook(WINDOW_COMMAND_LIST, "%d %d", 
-					winref, window ? window->refnum : -1);
+					winref, window ? (int)window->refnum : -1);
 				break;
 			}
 		}
